multi-organ seg/requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
opencv-python==4.1.1.26
protobuf==3.10.0
six==1.13.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
